requirements.txt
requests
python-dotenv
pycryptodome
tqdm
